+++ plugin/plugin/src/test/kotlin/net/twisterrob/gradle/vcs/SVNPluginIntgTest.kt
@@ -3,6 +3,7 @@ package net.twisterrob.gradle.vcs
 import net.twisterrob.gradle.BaseIntgTest
 import net.twisterrob.gradle.test.assertHasOutputLine
 import net.twisterrob.gradle.test.root
+import org.gradle.util.GradleVersion
 import org.intellij.lang.annotations.Language
 import org.junit.Test
 
@@ -20,12 +21,12 @@ class SVNPluginIntgTest : BaseIntgTest() {
 		@Language("gradle")
 		val script = """
 			apply plugin: 'net.twisterrob.vcs'
-			println(project.VCS.current)
+			println("VCS.current: " + project.VCS.current)
 		""".trimIndent()
 
 		val result = gradle.run(script).build()
 
-		result.assertHasOutputLine(""".*${SVNPluginExtension::class.qualifiedName}_Decorated@.*""".toRegex())
+		result.assertHasOutputLine("""VCS.current: extension '${SVNPluginExtension.NAME}'""".toRegex())
 	}
 
 	@Test fun `svn revision detected correctly`() {